requirements.txt
sounddevice
numpy
websocket-client
python-dotenv
vosk
colorama

